Application/requirements.txt
Flask==2.2.2
requests==2.28.1
flask-cors==3.0.10
beautifulsoup4==4.11.1
nltk==3.7
python-dotenv==0.19.0